F1Forecast/pyproject.toml
[project]
name = "Workspace_F1"
version = "0.1.0"
description = "AI-driven Formula 1 predictions with cinematic visuals and real-time data."
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.32.0",
    "gunicorn>=23.0.0",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
    "sift-stack-py>=0.8.4",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
]
